# internal/router/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Service Manager API",
    "version": "1.0.0",
    "description": "API Documentation for Service Manager service, including authentication, invites, password resets, and GitHub App integrations."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Input your JWT token in the format: `Bearer <token>`"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "unauthorized" }
        }
      },
      "SuccessMessageResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "example": "operation completed successfully" }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "admin@example.com" },
          "password": { "type": "string", "example": "AdminPassword123!" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email", "example": "admin@example.com" },
          "role": { "type": "string", "enum": ["admin", "user"], "example": "admin" },
          "github_installed": { "type": "boolean", "example": false },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-08-14T00:00:00Z" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
              "user": { "$ref": "#/components/schemas/UserResponse" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/users/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Verifies credentials and returns a JWT access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/users/": {
      "get": {
        "summary": "List all users",
        "description": "Returns a list of all registered users (Admin Only).",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UserResponse" }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        }
      }
    },
...
    "/users/{id}": {
      "delete": {
        "summary": "Delete a user account",
        "description": "Removes a user account from the system (Admin Only).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "User ID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessageResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}